requirements.txt
# Development
django-debug-toolbar>=1.10,<2.3
django-extensions>=2.1,<3.1
flake8>=3.8,<3.9
isort>=5.1,<5.2
django-modeltranslation

# Testing
coverage>=5.1,<5.3
django-webtest>=1.9,<1.10
factory-boy>=2.12,<2.13
pinocchio==0.4.2
psycopg2-binary>=2.8,<2.9
pytest-cov>=2.7,<2.11
pytest-django>=3.9,<3.10
tox>=3.1,<3.18
